debug_req: Move request association to `id_stage`
The `debug_req_i` signal is associated to a new exception value which is
in the private reserved space of RISC-V. So it should never collide with
future exception values, although this should be transparent to software
anyway.  Furthermore this patch makes the atomic commit logic simpler as
we are not getting any unexpected exceptions past the `id_stage`.

This patch is in preparation to fixing speculative reads to I/O regions.