Added directed tests for:
debug_req_i, irq and trigger on same cycle, with trigger insn being
 - ebreak
 - illegal instruction
 - branch instruction
 - multicycle instruction

Note that this is a fiddly piece of code in c and assembly,
and timing for the different events may change with compiler
and/or code changes. It may need a rewrite, but works for me now.

Signed-off-by: Oystein Knauserud <Oystein.Knauserud@silabs.com>