icache: Reset FSM to FLUSH (#668)
The current implementation resets to IDLE, where the icache checks if it
was just enabled (rising edge of en_i), and flushes (initializes) itself
if so. This assumes that en_i is de-asserted during reset, which is not
necessarily the case (e.g. hardwired `en_i`), possibly leaving the cache
uninitialized. Hence, reset to FLUSH directly.

Signed-off-by: Nils Wistoff <nwistoff@iis.ee.ethz.ch>