csr: Fix pmpaddr and perf counter privileges (#698)
* pmp: Fix csr

- Add proper granularity (G=1)
- Disallow NA4
- Fix read of pmpaddr
- Add PMP benchmark to CI

Signed-off-by: Moritz Schneider <moritz.schneider@inf.ethz.ch>

* Fix mcounteren/scounteren CSR logic

The privilege check should verify that the __current__ privilege level has
access to the counter CSRs. Before it checked if the privilege level of
the CSR instruction has access to the counters. Essentially, the m-mode
could not access any counter csr if mcounteren/scounteren was not set.

* Fix pmpcfg csr incase of NA4

The next pmpcfg register value needs to be set to the previous value if
NA4 is selected.