:bug: wb_dcache: Fix unaligned SC return data bug (#838)
When a store-conditional fails due to a missing reservation,
by specification it must return a non-zero value.
Previously a value of 64'b1 was returned by the axi_adapter.
However, when an atop's address is non-64bit-aligned, the return data
has to be realigned. This realignment causes the lower half-word
being lost, and the return datum taking a value of '0.

This commit fixes the bug by returning {64{1'b1}} from the axi_adapter.