Switch to GCC 13
Implementation of a new argument isa_extension in cva6.py to switch to gcc 13.1.0.
This argument includes the specific extensions to add to the isa for compilation.
It takes zicsr by default (extension to enable csr instructions).
Bugs fixed on: dv-riscv-compliance and dv-riscv-tests (need of enabling zifencei also), smoke_test (use of one linker instead of two).