ci: move expected results to cva6 repository
It makes it possible to perform changes which have an impact on CVA6
area without having to open a second MR on core-v-verif.

It is still needed to "declare" the change of area in a diff which is
reviewed, but now the diff is in the same repository as the
modification.