Document how to load and run zephyr on FPGA (#2949)
Starting with the upcoming release 4.2.0 (and the current main branch), the zephyr RTOS has support for running on the 32/64 bit configurations of cva6 on the Genesys2 board.
This PR adds instructions for building and loading zephyr for these configurations to the README.
It also adds a script that can be used to load zephyr from an SD card.
I also have automated scripts that setup the build dependencies for zephyr and build a sample application - I can add them to the PR and add a corresponding GitHub action if you are interested.
This closes #2480.

Co-authored-by: JeanRochCoulon <jean-roch.coulon@thalesgroup.com>