feat: Add support for SVNAPOT extension (#3094)
The CVA6 processor core lacked support for the RISC-V Svnapot (Naturally Aligned Power-of-Two) extension. The primary motivation for this feature is to improve MMU performance by allowing a single TLB entry to map a large, contiguous 64KiB memory region, thereby reducing TLB misses for memory-intensive applications. With this implementation Svnapot will be supported in CVA6 core.

Co-authored-by: JeanRochCoulon <jean-roch.coulon@thalesgroup.com>